Improvement1: We shouldn't analyse from the first audio frame of the audio 2 because the user may first pressed the Camera recording button and after few seconds, the Audio Recorder button is pressed then. So we make a pre-shift to the audio2, analysing a not-to-close-to-begining audio clip. If the child audio is shorter than 9min, it's 1/3 point is where the first analysing-frame sets. If the child audio is longer than 9 min, the 3min point is where the first analysing-frame sets.
Improvement2:
Suppose we set the trim to 15min.
The former BBC solution can only analyse the first 15min of the mother audio
But if the audio and video are very long, the actual offset may be longer than 15min, say: 30min.
In this circumstance, we may not be able to get the correct offset.
So, it is needed to cut the mother audio into slices, whose length is 15min with a pre-offset of 60s.
Then we compare the child frames to the mother-clips one by one, until we get a qualified score.

Improvement3:
Made it easier and more convinient to use.